+++ centroidDecomp.hpp
@@ -5,7 +5,7 @@
 #include <stdio.h>
 #include "allocUtils.hpp" // init*Vec, alloc*Vec, allocMat, copyMat
 
-void centroidDec( double **X, long n, long m, long truncated,
+void centroidDec( double **X, int n, int m, int truncated,
                   const char* matrixR, const char* matrixL,
                   std::ofstream &rmseFile);
 double** load_matrix(std::istream* is, int n, int m,
